move _is_pypi_release and _is_bentoml_in_develop_mode
